% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undumbify.R
\name{undumbify}
\alias{undumbify}
\title{FUNCTION_TITLE}
\usage{
undumbify(in_data, in_names)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{in_names}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
